Models, rm fixes. (#119)
- Downloading models page is easier. No fiddling with the model queue.
- Found every instance of `rm -r -f`. Guarded with a check that directory name is non-empty and is a directory. Should fix a problem I ran into where I unintentionally deleted some things on my Mmojo Share. Ooops.

-Brad

+++ scripts/mm-build-cosmopolitan.sh
@@ -13,7 +13,7 @@
 SCRIPT_NAME=$(basename -- "$0")
 printf "\n$STARS\n*\n* STARTED: $SCRIPT_NAME.\n*\n$STARS\n\n"
 
-if [ -d "$BUILD_COSMOPOLITAN_DIR" ]; then
+if [ "$BUILD_COSMOPOLITAN_DIR" != ""  ] && [ -d "$BUILD_COSMOPOLITAN_DIR" ]; then
     rm -r -f $BUILD_COSMOPOLITAN_DIR
 fi
