+++ scripts/unofficial/307-Clone-Build-llamafile.sh
@@ -13,7 +13,7 @@
 export BUILD_LLAMAFILE_DIR="$HOME/300-BUILD-llamafile"
 
 
-if [ -d "$BUILD_LLAMAFILE_DIR" ]; then
+if [ "$BUILD_LLAMAFILE_DIR" != "" ] && [ -d "$BUILD_LLAMAFILE_DIR" ]; then
     rm -r -f $BUILD_LLAMAFILE_DIR
 fi
